Remove links to individual test suites
Eliminates the need to create a link for every test suite, which
all point to either the 'legacy' or 'multirepo' suite.
This is accomplished by standardizing the naming convention for test
suites, and using a substring of the suite name to point to the correct
test script. This works as long as the test suite name ends in either
'-legacy' or '-multirepo'.